Remove two line navigation mappings
These mappings cause problems in nowrap mode by only sending
the cursor to the far left of the visible screen, rather than the
start of the line (this a problem for long lines that run off the
screen).

Initially I wanted the mappings to help in wrap mode with Goyo.vim,
but the pitfalls outweighed the benefits.